+++ src/Akeneo/AssetManager/back/Infrastructure/Transformation/Operation/ColorspaceOperationApplier.php
@@ -60,6 +60,7 @@ class ColorspaceOperationApplier implements OperationApplier
                     ],
                 ],
                 'quality' => 100,
+                'format' => 'png'
             ]
         );
         $this->filesystem->dumpFile($file->getRealPath(), $computedImage->getContent());