+++ src/Akeneo/AssetManager/back/Infrastructure/Transformation/Operation/ResolutionOperationApplier.php
@@ -62,6 +62,7 @@ class ResolutionOperationApplier implements OperationApplier
                     ],
                 ],
                 'quality' => 100,
+                'format' => 'png'
             ]
         );
         $this->filesystem->dumpFile($file->getRealPath(), $computedImage->getContent());