+++ src/Akeneo/ReferenceEntity/tests/back/Integration/Persistence/Sql/CLI/RefreshRecordLinksTest.php
@@ -94,7 +94,6 @@ class RefreshRecordLinksTest extends SqlIntegrationTestCase
         $commandTester = new CommandTester($command);
         $commandTester->execute([
             'command' => $command->getName(),
-            '--all'   => true,
         ]);
     }