+++ std-build/install-required-files.sh
@@ -2,15 +2,30 @@
 #
 # Script to be launched from the standard distribution to add missing files
 # add installation files.
-#
-# This script is only meant for installation from scratch, not for upgrade
-#
+
 
 set -e
 
 DEV_DISTRIB_DIR=$(dirname $0)/..
 STANDARD_DISTRIB_DIR=./
 
+
+# Step: Install and upgrade
+###########################
+
+# Update doctrine migratation files
+mkdir -p $STANDARD_DISTRIB_DIR/upgrades/
+cp -R $DEV_DISTRIB_DIR/../pim-community-dev/upgrades/* $STANDARD_DISTRIB_DIR/upgrades/
+cp -R $DEV_DISTRIB_DIR/upgrades/* $STANDARD_DISTRIB_DIR/upgrades/
+
+# Dot env file
+cp $DEV_DISTRIB_DIR/.env $STANDARD_DISTRIB_DIR/
+
+# Akeneo managed Kernel
+[ -d $STANDARD_DISTRIB_DIR/src ] && cp $DEV_DISTRIB_DIR/std-build/Kernel.php $STANDARD_DISTRIB_DIR/src
+
+# Step: Install
+###########################
 [ -d $STANDARD_DISTRIB_DIR/src ] && echo "src/ directory already exists. Not preparing the directory content." && exit 0
 
 # Required directories
@@ -18,8 +33,7 @@ mkdir -p $STANDARD_DISTRIB_DIR/src \
          $STANDARD_DISTRIB_DIR/bin \
          $STANDARD_DISTRIB_DIR/public \
          $STANDARD_DISTRIB_DIR/config/packages/dev \
-         $STANDARD_DISTRIB_DIR/config/packages/prod_flex \
-         $STANDARD_DISTRIB_DIR/config/packages/prod_onprem \
+         $STANDARD_DISTRIB_DIR/config/packages/prod \
          $STANDARD_DISTRIB_DIR/config/services \
          $STANDARD_DISTRIB_DIR/docker \
          $STANDARD_DISTRIB_DIR/docker/initdb.d
@@ -29,17 +43,14 @@ cp $DEV_DISTRIB_DIR/../pim-community-dev/docker/wait_docker_up.sh $STANDARD_DIST
 cp $DEV_DISTRIB_DIR/../pim-community-dev/docker/httpd.conf $STANDARD_DISTRIB_DIR/docker/
 cp $DEV_DISTRIB_DIR/../pim-community-dev/docker/akeneo.conf $STANDARD_DISTRIB_DIR/docker/
 
-# We use the same bootstrap.php to load .env file on standard as on CE-dev
+# We use the same bootstrap.php to load .env file on standard as on EE-dev
 cp $DEV_DISTRIB_DIR/config/bootstrap.php $STANDARD_DISTRIB_DIR/config/
 
-# Security configuration cannot be read on CE-dev and overriden in standard. We need to fully copy it
+# Security configuration cannot be read on EE-dev and overriden in standard. We need to fully copy it
 cp $DEV_DISTRIB_DIR/config/packages/security.yml $STANDARD_DISTRIB_DIR/config/packages/security.yml
 
 # Partners are most likely to develop and deploy using local filesystem, not MinIO
 cp -r $DEV_DISTRIB_DIR/config/packages/dev $STANDARD_DISTRIB_DIR/config/packages/
-
-# Both production server mode must be available in std
-cp -r $DEV_DISTRIB_DIR/config/packages/prod_* $STANDARD_DISTRIB_DIR/config/packages/
 
 # We need a console and FPM entrypoint
 cp $DEV_DISTRIB_DIR/bin/console $STANDARD_DISTRIB_DIR/bin/
@@ -69,13 +80,8 @@ cp $DEV_DISTRIB_DIR/std-build/tsconfig.json $STANDARD_DISTRIB_DIR/tsconfig.json
 cp $DEV_DISTRIB_DIR/std-build/services.yml $STANDARD_DISTRIB_DIR/config/services/
 
 # Skeleton .env file
-cp $DEV_DISTRIB_DIR/.env $STANDARD_DISTRIB_DIR/
 cp $DEV_DISTRIB_DIR/.env.dev $STANDARD_DISTRIB_DIR/
 
 # Skeleton .env file
 cp $DEV_DISTRIB_DIR/.gitignore $STANDARD_DISTRIB_DIR/
 
-# Prepare database upgrades to run
-mkdir -p $STANDARD_DISTRIB_DIR/upgrades/
-cp -R $DEV_DISTRIB_DIR/../pim-community-dev/upgrades/* $STANDARD_DISTRIB_DIR/upgrades/
-cp -R $DEV_DISTRIB_DIR/upgrades/* $STANDARD_DISTRIB_DIR/upgrades/